break out tests by contract
