+++ package.json
@@ -45,7 +45,6 @@
     "@types/mustache": "^4.2.1",
     "@typescript-eslint/eslint-plugin": "^2.0.0",
     "@typescript-eslint/parser": "^2.0.0",
-    "apollo-fetch": "^0.7.0",
     "assemblyscript": "0.19.23",
     "chai": "^4.3.6",
     "eslint": "^7.11.0",
