+++ subgraphs/venus-governance/tests/integration/constants.ts
@@ -1,0 +1,3 @@
+export const SUBGRAPH_ACCOUNT = 'venusprotocol';
+export const SUBGRAPH_NAME = 'venus-governance';
+export const SYNC_DELAY = 2000;
